🚚 Refactor tfcind test and move it to tfcoil test file; remove redundant code from SuperconductingTFCoil class
